Split draft screen into separate file
